refactor ui part of the updater to separate module
